backend/requirements.txt
setuptools
django
djangorestframework
djangorestframework-simplejwt
Pillow
django-filter